pyproject.toml
[tool.poetry]
name = "labelizer"
version = "0.2.0"
description = "A protein analysis toolbox for fluorescent labelling and FRET assays"
authors = ["Christian Gebhardt <chrirstian.gebhardt@bio.lmu.de>", "Pascal Bawidamann", "Konstantin Schütze <schuetze@in.tum.de>"]
readme = "Readme.md"
license = "MIT plus https://ccsb.scripps.edu/msms/license/ for MSMS"

[tool.poetry.dependencies]
python = "^3.8"
biopython = "^1.7"
pandas = "^1.5.2"
scipy = "^1.9.3"
labellib = "2020.10.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
